repo_content/telegram_bot/requirements.txt
python-telegram-bot[webhooks]==21.7
httpx>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
pydub>=0.25.1
PyPDF2>=3.0.0
python-docx>=1.1.0
aiofiles>=23.2.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
respx
gtts>=2.5.0
pdfplumber>=0.11.0
Pillow>=10.4.0
structlog>=24.1.0
aiosqlite>=0.20.0
